Solved: Largest Odd Number in String (#2032)

+++ problems/2032-largest-odd-number-in-string/solution.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     string largestOddNumber(string x) {
-        int num = stoi(x);
+        long long num = stoll(x);
         while(num > 0 && num%2 == 0){
         int last = num%10;
         num = num -  last;
